+++ release-notes.md
@@ -1,3 +1,11 @@
+`0.1.8` - ?
+* Added the ability to set attack animations per item ([#31](https://github.com/wrycu/StarWarsFFG-Enhancements/issues/31))
+* Improve dependency management for JB2A Patreon module ([#27](https://github.com/wrycu/StarWarsFFG-Enhancements/issues/27))
+* Fix for attack animations only working on English version ([#34](https://github.com/wrycu/StarWarsFFG-Enhancements/issues/34))
+* Fix for dice helper only working on English version ([#34](https://github.com/wrycu/StarWarsFFG-Enhancements/issues/34))
+* Fix for shop generator button not being localized ([#36](https://github.com/wrycu/StarWarsFFG-Enhancements/issues/36))
+* Fix for `Mandar` theme mucking up actor sheet names ([#37](https://github.com/wrycu/StarWarsFFG-Enhancements/issues/37))
+
 `0.1.7` - 2021-05-18
 * Initial alpha release
 * Planned features: